refactor(claude-appserver): extract session contract and env helper from controller
Pulls the broker session vocabulary and env helpers out of the
875-LOC controller so server code stops depending on the controller
implementation file.

- `session-contract.ts` (NEW) owns broker session types:
  `ControllerNotificationMap`, `ControllerNotification`,
  `ClaudeBrokerChild`, `ClaudeBrokerSession`,
  `CreateBrokerSessionOptions`, `SpawnClaudeChildOptions`,
  `SingleSessionControllerOptions`, plus `ChildExit` (which travels
  with `ClaudeBrokerChild.onExit`). All controller-private types
  (`ChildBinding`, the SDK message variants, `ActiveTurnState`,
  `PendingControlRequest`) stay inside `controller.ts`.
- `child-env.ts` (NEW) owns `buildClaudeChildEnv` only. The hash
  step is `hashSortedEnv` (already exported from
  `providers/claude/request-prep.ts`); a `hashClaudeChildEnv`
  rename was a pure pass-through and got dropped — callers
  compose `hashSortedEnv(buildClaudeChildEnv(...))` directly.
- `controller.ts` re-imports both. The class body, `stderrRing`,
  and `appendRingBuffer` stay in place.
- `server.ts` imports `buildClaudeChildEnv` from `child-env.js`.
  `broker-pool.ts` imports session vocabulary from
  `session-contract.js`; its bottom re-export now sources from the
  same file.
- Unit tests updated to import from the new sibling modules.
