+++ src/providers/claude-appserver/child-env.ts
@@ -1,0 +1,6 @@
+import { buildChildEnv } from '../../infra/env-sanitize.js';
+import { normalizeControllerEnv } from '../claude/request-prep.js';
+
+export function buildClaudeChildEnv(controllerEnv?: Record<string, string>): Record<string, string> {
+  return buildChildEnv(normalizeControllerEnv(controllerEnv));
+}
